little refactoring
