+++ easy_manage/easy_manage.py
@@ -1,6 +1,7 @@
 import argparse
-import utils
 import pprint
+import utils
+import redfish_controller
 
 
 def parse_args():
@@ -19,7 +20,7 @@ def parse_args():
 def main():
     args = parse_args()
 
-    testfish = utils.RedfishController('testfish', args.address, args.port)
+    testfish = redfish_controller.RedfishController('testfish', args.address, args.port)
     test_sys = testfish.systems[0]
     print(test_sys)
     testfish.data = testfish.update_recurse('/redfish/v1/Systems/System-1')
